requirements.txt
discord.py
python-dotenv
spotipy
pytube